Separate the functions for LLB creation
Move the functions related to the construction of the LLB to a new file.
